Share PageNavigation (bottom paging) component and add it to docs pages
Extract the bottom Previous/Next page navigation from the UI site into
a shared component with inline SVGs, and render it at the bottom of
every docs page.
